requirements.txt
streamlit==1.32.2
rich<14,>=10.14.0
groq==0.5.0
requests==2.31.0
lingua-language-detector==2.0.2
Pillow==10.3.0
pytesseract==0.3.10
python-docx==1.1.2
pymupdf==1.24.3
numpy==1.26.4
pyarrow==15.0.2
protobuf>=3.20,<5
gtts==2.4.0
pandas==2.1.4
plotly==5.18.0
streamlit-feedback==0.1.3
supabase==2.1.5